Add files via upload
This is the first image viewer I have created using Tkinter.
This is not so much as an effective program. Not even close.  For the moment, I am not putting much effort into the logic 
of the program but rather exploring the GUI Framework of Python, Tkinter. I am uploading these projects to gauge my own progress

As of right now, the program will require the uploaded images and .ico file to be in the same directory as the .py script.
There are several inconsistencies, but I chose the pictures of the same resolution to prevent the window from changing its own size on every picture.
Note: To display jpg files, Pillow is required as mentioned in the comments.